// src/typing/enums/Colors.ts
enum Colors {
	color_basic_1 = '#544B95',
	color_basic_2 = '#6556AA',
	color_basic_3 = '#D8D4ED',
	color_basic_4 = '#F1F0FD',
	color_basic_5 = '#FDEEB6',
	color_basic_6 = '#FBDE75',

	color_white_1 = '#FFF',
	color_white_2 = '#F9F9FA',

	color_black_1 = '#000',
	color_black_2 = '#4D4D4D',

	color_gray_1 = '#C2C2C2',

	color_error = '#D45057',

	color_status_badge_pending = 'rgba(255, 206, 86, 1)',
	color_status_badge_pending_bg = 'rgba(255, 206, 86, 0.2)',
	color_status_badge_approved = 'rgba(75, 192, 192, 1)',
	color_status_badge_approved_bg = 'rgba(75, 192, 192, 0.2)',
	color_status_badge_rejected = 'rgba(255, 99, 132, 1)',
	color_status_badge_rejected_bg = 'rgba(255, 99, 132, 0.2)',

	color_diagram_1 = ' rgba(255, 99, 132, 1)',
	color_diagram_1_bg = 'rgba(255, 99, 132, 0.2)',
	color_diagram_2 = 'rgba(255, 206, 86, 1)',
	color_diagram_2_bg = 'rgba(255, 206, 86, 0.2)',
	color_diagram_3 = 'rgba(153, 102, 255, 1)',
	color_diagram_3_bg = 'rgba(153, 102, 255, 0.2)',
	color_diagram_4 = 'rgba(75, 192, 192, 1)',
	color_diagram_4_bg = 'rgba(75, 192, 192, 0.2)',
	color_diagram_5 = 'rgba(54, 162, 235, 1)',
	color_diagram_5_bg = 'rgba(54, 162, 235, 0.2)',
}

export default Colors;