// src/typing/enums/IconNames.ts
enum IconNames {
	applications = 'applications',
	check = 'check',
	cross = 'cross',
	job = 'job',
	messages = 'messages',
	notifications = 'notifications',
	profile = 'profile',
	people = 'people',
	search = 'search',
	settings = 'settings',
	signOut = 'signOut',
	statistics = 'statistics',
	arrow = 'arrow',
}

export default IconNames;